+++ src/hydromt_fiat/components/__init__.py
@@ -1,0 +1,5 @@
+"""The custom HydroMT-FIAT components."""
+
+from .hazard import *
+from .region import *
+from .vulnerability import *
