+++ tests/components/conftest.py
@@ -1,0 +1,22 @@
+from unittest.mock import MagicMock, PropertyMock
+
+import pytest
+from hydromt import DataCatalog
+from hydromt.model import ModelRoot
+from pyproj.crs import CRS
+from pytest_mock import MockerFixture
+
+from hydromt_fiat import FIATModel
+
+
+## Mocked objects
+@pytest.fixture
+def mock_model(tmp_path, mocker: MockerFixture) -> MagicMock:
+    model = mocker.create_autospec(FIATModel)
+    model.root = mocker.create_autospec(ModelRoot(tmp_path), instance=True)
+    model.root.path.return_value = tmp_path
+    model.data_catalog = mocker.create_autospec(DataCatalog)
+    # Set attributes for practical use
+    type(model).crs = PropertyMock(side_effect=lambda: CRS.from_epsg(4326))
+    type(model).root = PropertyMock(side_effect=lambda: ModelRoot(tmp_path))
+    return model
