+++ tests/components/test_vulnerability_component.py
@@ -1,0 +1,36 @@
+from hydromt_fiat.components import VulnerabilityComponent
+
+
+def test_vulnerability_component_empty(
+    mock_model,
+):
+    # Setup the component
+    component = VulnerabilityComponent(model=mock_model)
+
+    # Assert the content
+    assert component._filename == "vulnerability/{name}.csv"
+    assert len(component.data) == 0
+    assert isinstance(component.data, dict)
+
+
+def test_vulnerability_component_setup(model):
+    # Setup the component
+    component = VulnerabilityComponent(model=model)
+
+    # Assert it's empty
+    assert len(component.data) == 0
+
+    # Setup the vulnerability
+    component.setup_vulnerability(
+        vulnerability_fname="jrc_vulnerability_curves",
+        vulnerability_linking_fname="jrc_vulnerability_curves_linking",
+        continent="europe",
+    )
+
+    assert len(component.data) == 2
+    assert "vulnerability_curves" in component.data
+    assert "vulnerability_identifiers" in component.data
+    assert (
+        model.config.get_value("vulnerability.file")
+        == "vulnerability/vulnerability_curves.csv"
+    )
