+++ tests/test_model.py
@@ -81,83 +81,6 @@ def test_setup_region_error(tmp_path):
         model.setup_region(region=region_no)
 
 
-def test_setup_vulnerability(tmp_path, build_data_catalog):
-    # Setup the model
-    model = FIATModel(tmp_path, mode="w+", data_libs=build_data_catalog)
-
-    assert len(model.vulnerability_data.data) == 0
-
-    # Setup the vulnerability
-    model.setup_vulnerability(
-        vulnerability_fname="jrc_vulnerability_curves",
-        vulnerability_linking_fname="jrc_vulnerability_curves_linking",
-        continent="europe",
-    )
-
-    assert len(model.vulnerability_data.data) == 2
-    assert "vulnerability_curves" in model.vulnerability_data.data
-    assert "vulnerability_identifiers" in model.vulnerability_data.data
-    assert (
-        model.config.get_value("vulnerability.file")
-        == "vulnerability/vulnerability_curves.csv"
-    )
-
-
-def test_setup_hazard(model, tmp_path, build_data_catalog, caplog, build_region):
-    # Setup the model
-    model.setup_region(region=build_region)
-    # Test hazard event
-    caplog.set_level(logging.INFO)
-    model.setup_hazard(hazard_fnames="flood_event")
-
-    assert "Added flooding hazard map: flood_event" in caplog.text
-    assert model.config.get_value("hazard.file") == "hazard/hazard_grid.nc"
-    assert model.config.get_value("hazard.elevation_reference") == "datum"
-
-    # Test setting data to hazard grid with data
-    model.setup_hazard(hazard_fnames="flood_event_highres")
-    assert model.config.get_value("hazard.settings.var_as_band")
-
-    # Check if both ds are still there
-    assert "flood_event" in model.hazard_grid.data.data_vars.keys()
-    assert "flood_event_highres" in model.hazard_grid.data.data_vars.keys()
-
-    # Test hazard with return period
-    model2 = FIATModel(tmp_path, data_libs=[build_data_catalog])
-    model2.setup_region(region=build_region)
-    model2.setup_hazard(
-        hazard_fnames=["flood_event_highres"],
-        risk=True,
-        return_periods=[50000],
-    )
-
-    assert isinstance(model2.hazard_grid.data, xr.Dataset)
-    assert model2.config.get_value("hazard.risk")
-    assert model2.config.get_value("hazard.return_periods") == [50000]
-
-
-def test_setup_hazard_errors(model):
-    with pytest.raises(
-        ValueError, match="Cannot perform risk analysis without return periods"
-    ):
-        model.setup_hazard(hazard_fnames="test.nc", risk=True)
-
-    with pytest.raises(
-        ValueError, match="Return periods do not match the number of hazard files"
-    ):
-        model.setup_hazard(
-            hazard_fnames=["test1.nc", "test2.nc"],
-            risk=True,
-            return_periods=[1, 2, 3],
-        )
-
-    with pytest.raises(
-        MissingRegionError,
-        match=("Region component is missing for setting up hazard data."),
-    ):
-        model.setup_hazard(hazard_fnames=["flood_event"])
-
-
 def test_setup_exposure_grid(model, build_region, caplog, tmp_path, mocker):
     model.setup_region(region=build_region)
